requirements.txt
ultralytics==8.0.0
torch==2.0.0
torchvision==0.15.0
opencv-python==4.8.0.74
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0
tqdm==4.65.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
scipy==1.11.1
scikit-learn==1.3.0
